backend/requirements.txt
asgiref==3.11.0
black==25.11.0
boto3==1.42.4
botocore==1.42.4
click==8.3.1
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==23.0.0
jmespath==1.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytokens==0.3.0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
urllib3==2.6.0
whitenoise==6.11.0
